requirements.txt
asgiref==3.5.2
Django==3.2
django-filter==21.1
djangorestframework==3.11.0
djangorestframework-simplejwt==5.1.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2
whitenoise==6.1.0
